compare.py: Implement parallel queries between servers
Now, instead of querying each server serially, both connections and
SELECTS happen in parallel between servers, so they only wait for
themselves, and not to each other.

Other improvements such as print-every option and output formatting.

Now error codes are working: returns 0 if tables are identical,
1 if the comparison finished but there are differences, and -1 if
there was an error or the script was aborted.

Change-Id: I5231ea606e715e9f3e53b3caff58fe30db7596a6