mariadb: Added functionality to perform arbitrary topology changes
A first version of a non-trivial topology change between a replica
and another host has been implemented.
This case requires that the instance is replicating from a server,
it is active and has a small amount of lag. The target host also
has to have very little lag or be a master (not replicating).

Many assumptions are used on this first version- it is based on
GTID serial numbers -which may not be working in all cases (e.g.
side effects), however changing it to a different implementation,
such as heartbeat or another type of pseudo-gtid could reuse most
of the code.

In this first implementation, however, there is very little
optimization -it scans full binary logs at a time on memory, which
can be a killer of performance. This has to be tuned before being
useful in production-sized servers, with gigabytes of data.

Change-Id: Iec8c04b8fca9750bc1672c251840880497e91986