WMFReplication: Fix bug where the wrong address was chosen on move()
The current implementation, where GTID position is used caused a
duplicate key, as it tried to execute the same transaction twice.
Now the same transaction is tried, but then we try to find the next
GTID in order.

Code will be need to be refactored, highly simplified and we should
implement something not based on gtid. However, it should work for
the most cases for now (assuming no GTID bugs- which is not really
a huge guarantee).

Also, to speed up the process, `set sql_log_bin =0; flush binary logs`
should be run on the replica and new master to speed up the binary
log reading process, as a temporary workaround.

Change-Id: Ia5d2625531606a4082b7156628e5b49fd80e8a73