switchover: Make posible replica migration an optional, separate step
Right now, switchover performs a replica migration before any other
change on the current master (it moves all direct replicas to
replicate from the new master). Introduce 2 new parameters,
--skip-slave-move, which skips that step (so it can be done
beforehand or afterwards), and --only-slave-move, which only
performs the checks and the replica migration and then exits.

This way the migration, which is the weakest part of this script
can be done with a separate method or at a separate time than
the rest of the steps.

Also document all parameters so it is clear what it is the
expected functionality and options.

Bug: T199224
Change-Id: I8014100b81d3c551b9c4d53435e38921b55e2c02