switchover: Add the functionality to start and stop heartbeat
This has been tested, and it will kill and start
pt-heartbeat-wikimedia automatically on failover. To avoid cumin
being stuck on start (running as a daemon) we run pt-heartbeat
with nohup '&> /dev/null &'.

It uses RemoteExecution with Cumin bindings for remote execution
to handle heartbeat (running locally at the moment on each
db host).

Make it optional so that it we don't depend on it (there may be
hosts where there is no heartbeat, or we may handle that
separatedly).

Bug: T199224
Change-Id: Ib449a64a21a2c835464d3dc729c0274e4d303bb1