mariadb: Fix heartbeat regex
The heartbeat execution that is active on some hosts do not match
that of what it is on puppet. This create some issues with
heartbeat not being able to be detected/stopped.

While it would be trivial to detect any kind of heartbeat
execution with a looser detection, we actually need to parse the
parameters, so we should check the run is updated to the latest
puppet code on all hosts instead.

Bug: T199224
Change-Id: I111eeff19e600168a3510b5940283a70941d366d