switchover: Fix bug where shards are added with an extra 'b'
In situations where a running pt-heartbeat was not detected,
and this was detected from the heartbeat table, the section
was added as a paramater with an extra b at the start (e.g. bs1),
due to wrong decoding of binary strings.

They are decoded to ascii now to avoid the wrong transformation.

Bug: T199224
Change-Id: I9968a4bc42e0061fd58c157a2efc747b67df2613