transfer.py: Add the ability to transfer from a new mariabackup
New option, --type= { file | xtrabackup } is implemented. File
is the old behaviour (copy a file or sync a directory).

xtrabackup uses mariabackup to start a hot* backup process and
send it remotelly to the target host. In the case the xtrabackup
option is used, the source path should point to the socket of
the source database.

Some extra changes were required to implement this- socket
identification, new commands (mariabackup does not support tar,
it has to use xbstream) and new checks. Some code has been
refactored slighly to prevent code duplication.

Bug: T210292
Change-Id: Ia262ec5b8ee2879f4bcd1918b8fc6c641a82c529