mariadb: Add the option of postprocessing backups
Allow postprocessing of xtrabackup:

* It runs xtrabackup --prepare
* It gathers statistics, recursivelly
* It compresses the output on a final tar, in parallel (requires pigz)

Introduces a new option --only-postprocess, which skips the
actual taking of both dumps and snapshots, but searches for one
matching ongoing backup and does all the backup steps except
the actual mydumper or xtrabackup call. Useful if the data gathering
failed or it was done remotelly (most useful for xtrabackup steams).

It also changes how --archive option works:

* It is allowed now on command line (may need further work for
  refinement)
* On dumps, it just creates a tar with each database
* On snapshots, it compresses the whole backup in a single tar.gz
  with pigz (new dependency)

Bug: T210292
Change-Id: I6ea9496d2d65dc07256eb524042862264ee1f1e5