BackupStatistics: Narrow the search for the treated backup
We don't use last_insert_id(), in case it is a non-inserted on
the same session backup. Make the search for the last backup
narrower so it only matches the backup that is ongoing for the
current type, source, section, and within the last 24 hours.

Refactor into find_backup_id so we don't have redundant code.

Bug: T206203
Change-Id: Iac4d690c15c97ab5ec2608bb278dfed3d96a814c