WMFBackup: Reimplement os.rename to make it work on different fs
os.rename only works for a real, link-based rename (meaning it
only works if origin and destination are on the same device).

We want to play with different SSD and HD disks, so we need to
transparently handling that, rather than choosing to use os.rename
or shutil.copydir, etc. Execute /bin/mv instead.

Additonally, implement return code checking for both mv and tar, so
in case it fails, the backup is marked as failed.

Bug: T206203
Change-Id: Iee04e51ad03071a21bdc6817ef3b9c524affb6f4