transfer.py: Allow for a 3rd transfer type: decompression
Decompression is thought for sending a tar.gz of a snapshot
(backup generated with xtrabackup/mariabackup) and decompressing
it on-the-fly on the destination host.

Source will be the host and tarball file absolute path and
destination the empty target host + dir. If the target directory
is not empty, it will abort. To check the transfer has been
completed successfuly, it will check for an xtrabackup_info file.

Note that tarballs from backup are created with a directory like:
snapshot.s1.2020-09-11--15:17:23/ That one is stripped, so the
target dir should be /srv/sqldir or equivalent.

Example execution:
./transfer.py --type=decompress --no-checksum --no-encrypt \
dbprov2001.codfw.wmnet:/srv/backups/snapshots/archive/snapshot.x1.2019-03-28--12-56-40.tar.gz \
es2002.codfw.wmnet:/srv/tmp/sqldata/

Bug: T219631
Change-Id: I09554b2fb1c48baecec3e2a0b8f3bac1f650f147