mariadb: Allow new option --stop-slave for xtrabackup transfers
We belive that on high traffic databases, stopping the slave
before starting a backup could speed up substantially the process.
Create a new option for transfers, `--stop-slave`, which only
affects xtrabackup type copies, that stops the main replication
channel of a mariadb instance before backup starts, and starts it
later.

No check is being done if the database was already stopped, or
more complex setup, as this is supposed to be run only on
non-production hosts, and at some point it should be handled by
the WMFReplication API.

Bug: T206203
Change-Id: Ice5befa27e71bbc7935018a9d737cdb13a88a040