mariadb: Allow the passing of a full path to section on prepare
In case an --only-prepare run is done, allow to pass a full
path of the directory to be prepared instead of only the section.
This will prevent failures to cascade by indicating which
specific backup wants to be prepared, rather than failing because
it is ambiguous (there is a failed and a new one).

In other words, in addition to (which is still valid):

backup_mariadb.py s1 --only-postprocess [...]

you can now run:

backup_mariadb.py /tmp/snapshot.s1.2019-06-07--12-15-24 --only-postprocess [...]

Bug: T206203
Change-Id: I6044aff3925c4bf69324e7c27ac47aeef653872f