Add 2 simple scripts: move_replica.py and stop_in_sync.py
move_replica.py makes a replication topology change of moving the
database on the first parameter under the instance of the second
parameter. Note for now this only works for direct replica
changes, and not arbitrary locations.

stop_in_sync.py stops 2 replicas that are siblings in the same
exact coordinate, useful if some comparison or maintenance is
going to be done.

Change-Id: I58104d4d7587d09401b7022fd2b4ad8bea876afb