replication_tree.py: Console output of a replica set
Add script to output a text-based representation of a replica set,
given the name of the "master" of the replication (doesn't have
to be a real master, just it will be where it start the scan
for replicas. In the future it could be extended to show its
masters, too.

It output in a tree style the list of replicas along with some
(hopefully) useful information:
* Uptime
* read only state
* binlog format
* lag, if it is a replica
* number of non-system processes
* latency of an example query
* server version
* It marks with a different color cross-dc replication

It is for now not compatible with multi-source replication. QPS
is not displayed for now as it would slow down the data retrival.
Another improvement would be to add a top-style refresh, which would
allow to add a QPS counter.
A JSON output would also be nice as a libary.

As a side notice, WMFMariaDB was updated to fix a bug when
connecting to labsdb hosts (it was still using a deprecated key).
Also, it was added a parameter to name() to avoid showing the
database between brackets when it is not relevant.

This is too slow, requires WMFReplication.slaves() parallelization
for a nice a speedup (will also speedup switchover.py).

Change-Id: I8ca27daeb00b3f236bab376810674be3abcb45e9