switchover.py: Wait a few seconds after gtid disable to catch up
If a host is loaded and stop/start slave is slow, the replica may
have some lag after disabling gtid, making topology changes fail.
Add a pause after gtid disabling and before moving the replica
so that errors are less likely, even if topology changes are
slower, specially for large number of replicas (it can be done
in advance).

Change-Id: I440fbdeac8d4cb932774fd89ab0c5442e658e192