CuminExecution: Capture Exception cumin.transports.WorkerError
If a wrong host was given to transfer.py, an empty NodeSet will be returned.
An empty host will cause WorkerError which should be caught. In this case, the
script should be aborted gracefully, rather than raise Exception to the user.

Determine the returned hosts are empty or not. Return 1 as return code, None as
standard out, 'host is wrong or does not match rules' as standard error if hosts
are empty. Add a test case of a wrong host not in the known hosts. Reformat
imports with isort.

Bug: T218189
Change-Id: I7008570e15081ff777a14e981717793b48d1ccf0