transfer.py: Convert return for run() from int to list
The return code of the library (not the class) of transfer.py was not working as intended. The
intended behavior is to fix the cumin misleading behavior to not return 0 if the execution went
well, even if one of the remote executions went badly. To do this, the intention was to store
the return codes in a list and return the max of them.

The main function expects that the run function of Transferer class returns a list. This list is
then passed to the max function to get the integer with the maximum value from this list. A few
of the exception-handling blocks inside the run function returned integers (-1,-2,-3) instead of
a list which resulted in the max function to break and give a TypeError since it expected a list
and not an integer as a parameter.

Change the exception-handling blocks that return integers to the singleton lists
[-1], [-2] or [-3].

Unit tests to test whether the run function of the Transferer class returns a list under any
exception block.

Bug: T248661
Change-Id: Ib001d371d3cdae01c7b336b62d1bb7415c6e19e1