transfer.py: Let netcat's listen instance timeout after 10 seconds
Sometimes, if client fails to start a connection to send data
over the network, nc keeps listening forever (actually, the
tcp port expires and does not listen anymore but the command
keeps executing).

Make it expire after 10 seconds, so the execution finishes.
I tested and debian version of nc command applies also to
the listener.

Increase client (sender) wait to 3 seconds, in case the
server is too busy to start listening.

Change-Id: I7634e3b70c6e9b74dd1dd12faf79072a5b723e82