Firewall.py: Add function to kill process by its port number
This framework uses remote execution module's kill_job function to
kill the netcat listen process. But kill_job does not kill the
remote process instantly. This patch implements two new function
kill_process and find_pid for futrure use in the Firewall class.
kill_process is to make an instant kill for the port-using-process
and find_pid is to find the PID associated with a port number.

This patch uses fuser utility to find the PID associated with a port.

Bug: T252950
Change-Id: Ib662bc4975df2758c3c402587133150d85051437