mariabackup: Stop harcoding xtrabackup path
xtrabackup path was hard-coded to the binary on wmf-mariadb101
(mariabackup). Use the global path now that every package has
been upgraded, fixing the leftover TODO.

This will unblock using xtrabackup (mariabackup) from MariaDB 10.4,
as well as the generic xtrabackup package, so it works with other
installs that are not using the wmf-mariadb101 package.

Also fix some related documentation.

Bug: T250666
Change-Id: I88dc9e58b5e25397a232f931e442152f87685144