mariadb-backups: Add _ to the list of characters alowed for section names
The underscores were not considered valid section identifiers.
Because there was a limitation on lack of escaping on systemd
units for hyphens (-) on analytics dbs, an underscore was prefered,
leading to the addition of it to the list of valid sections,
used to rotate old backups.

This will solve issue of analytics_meta backups not being properly
rotated/purged.

Bug: T234826
Change-Id: I0b56436905c2f135cf0a427ce9e0b234c276e934