wmfbackups: Copy backup-related scripts from puppet to wmfmariadbpy
The following 2 scripts used to be on puppet only.

Copy them to the wmfbackups module of the wmfmariadbpy repo in
order to maintain them together. They will be eventually (along with
other redundancies) removed from puppet repo.

Fix to use the WMFMariaDB paths and functionalities.

Requires a minor refactoring of a static method of WMFMariaDB class.

Bug: T165358
Change-Id: I91147408d3bd1f088984aa76857a0f52ec4ac674