wmfmariadbpy: Load and provide a method for section to port assignment
MariaBackup required a both socket and datadir from a given port.

Create a configuration file to provide that match (in both ways),
and use it to load it from the filesystem on request and provide
several methods to query it. This mapping file could be provided
by the package or by puppet.

Update MariaBackup to use such method.

Bug: T165358
Bug: T257033
Change-Id: I6cc99b5ed75b83498cacbd1e963c2f0b224493a8