Add 'black' formatter support.
Black is only available for python3.6+, and we still want to to be able
to run the other envs on python3.5 (for debian stretch support). So this
change adds a python3.7 (the default version on debian buster) venv
specifically for 'black' (and black is run with '--target-version py35'
to enforce that changes are py3.5-compatible).

The next CR will format all wmfmariadbpy python files using black,
and the final CR will enforce black in CI.

Bug: T211750
Change-Id: Ife8a8fdf8ebe397e49a3d0f8f509ed0e775fa9f1