mariadb-backups: Fix missing check on no ERROR msgs on logs
Despite being coded, the logic for checking error messages on logs
was unused. This allowed errors on es4 and es5 to go on, as the
metadata files was anyway generated correctly. Fix this gap
and make sure dumps logs are checked before considering the
backup successful.

Also alter slightly the signature of the method.

Bug: T138562
Change-Id: I35cf13a6b367108be824afd2ccbee574a77eb9ab