remote backups: Modify the scheduler so it parallelizes per destination
In the previous scheduler, backups were executed by order and
with a parallelization of 2 (2 backups could run at the same time).
Now, with the introduction of a 3rd hosts, this is seen as even more
inefficient, specially for highly unbalanced distribution of backups,
as one destination host may be idle waiting for the backup to finish
on 2 or even just 1 other host.

Create 1 parallel process per destination, and execute them
independently, so all are busy at the same time. Execute each backup
job within a destination serially, so it would have 100% of the
available resources.

This should speed up considerably the backup termination.

Additionally, don't say backups terminated successfully if one
fails, indicate show on final message.

Bug: T257551
Change-Id: Ibb56e6d64aa59ff72beb1c484c23b7d38835270d