Update patch set 2
Patch Set 2:

(1 comment)

Thanks for taking the time to fix that. This will fail until the other is merged.

Let's do also a scan of puppet and the transfer.py repo to make sure nothing is missing.

Let me know when you intend to deploy to deploy at the same time.

Patch-set: 2
CC: Gerrit User 2556 <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>