Update patch set 3
Patch Set 3:

So I am ready to merge, like this (the correct way) or with the git url (so it works on CI). Ready to deploy when you are, kormat :-D I have the packages ready for both OSes.

Patch-set: 3
Reviewer: Gerrit User 2556 <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>