Add sql with the empty database structure to the repo
The sql data was handled as data, and hence not part of the
repo code. At WMF, this was just backed up and recovered by the
puppet setup. This is not something that would work out of
the WMF installation, so adding it to the code repo.

It is only 3 tables, of which 1 is currently not used, that
track the backups performed and the files created on each backup.

This was documented already at
https://wikitech.wikimedia.org/wiki/MariaDB/Backups#Metadata
but it was not usable as code.

Bug: T138562
Change-Id: I201b7e9a8846e094ddbbaa17e2c1d69665f4e094