Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> If you don't mind some criticism, some of your changes look a bit too verbose. While clarity over brevity is usually preferred, something like:
> 
> recover_logical_dump_logger.info(msg="Decompressing {} ...".format(backup_name))
> 
> Seem too much. You can name the variable logger- we will be logging to the same location throughout the logic, so probably there will be no confussion.
> msg is normally implied, and format is not necessary if you look at the API documentation. Without using fstrings (because we need to support older versions of python3, I would write that as:
> 
> logger.info('Decompressing %s ...', backup_name)
> 
> I think it looks much cleaner. Just an opinion.

Sure, no worries - criticism is good feedback. 

I would say it's a habit (bad or good?) of mine to name things like that, but in this case I think you make a good point and it's better to be pragmatic than precise and rename it to logger just to make it cleaner, and as you said we are logging within the same program/location - so this shouldn't cause any issues if we keep the name as logging. Happy to do that in the next patch set.

Regarding fstrings, I wasn't aware of the old version requirements. The logger messages were made from the existing print statements which uses f-strings rather than %s strings. Example

print('Unarchiving {} ...'.format(db_tar_name))

There are perhaps 6 or 7 instances of f-string usages in the existing file.
I can change the loggers to use %s if that's preferred, I could also do the changes on print statement - or I could raise a ticket so that there are more tasks for potential GSoC applicants to practise contributing to projects.

For the next patchset, I could change the logger only messages to use %s strings rather than f-strings (unless you want me to do the same for the existing print statements as well - or unless it's okay to keep using f-strings)
Please let me know :)

Patch-set: 2