Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I am also unsure about the test- it only checks that it returns 0. Do you think it would be possible to check something else, like logging in a particular format, by overriding the logger method, or checking no exceptions are rised? Otherwise it may be of limited usefulness.

Yep that's a good point, I will have a look today and push in a patchset later tonight
Regarding my earlier comment - I will change the logs to %s-strings - however the current code base already has f-strings for the print statement. Are you happy for me to raise a Phabricator ticket later on as a micro task for someone else to pick up (and change the print statement as well?). I suppose if we are targetting Python 3.5 and below, we need to get rid of the f-strings as well?

Patch-set: 2