Update patch set 2
Patch Set 2:

> > > the current code base already has f-strings for the print statement

Apologies, I think there is a misscommunication on my side (my fault). When I mentioned fstrings, I meant literally f-strings (the syntax: f'string'): https://www.python.org/dev/peps/pep-0498/#specification

This is a replacement for format(). But before Pyhton 3.16 there were libraries that allowed other kinds of interpolation, such as mysql.query() and logger (both using %s).

We should avoid .format() when possible, but it is ok to keep using it (like on the places you mention) if there is no other option until we can target Python >3.16, in a few months.

There can be a reasonable reason to transform those into logger.info() calls, when reasonable, but that is a separate task. Not related for f-string adoption- as we still have to support for a few extra weeks Debian 9 (Pyhton 3.15).

For a new repo/new application, you will be able to use (and I recommend) fstrings, as we will not have to have backwards compatibility. Apologies for the confusion caused.

Patch-set: 2
Reviewer: Gerrit User 2556 <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>