Update patch set 3
Patch Set 3:

Yes it is clear now, sorry, I was also confused as well, too much haste to reply and I should have read that properly beforehand.
Thanks for your review and advice, I have added unit tests which actually writes to a log file and reads from the file and checks if the log is present. I had to add a new requirement to test, which is called freezegun - it helps unit tests to freeze time, and then check if the time appears in the logs correctly. 
I also noticed that when running nose tests through tox, nose modifies the test logger configuration, so I had to add a small override to tox.ini. A possible alternative it to use python -m discover unittests instead, skipping nose tests. (but that may not be ideal?)

Additionally, as mentioned previously, I've made the logger name in the recover-dump simpler. I have kept the log print statements to use .format(). This is just to keep it consistent with the print statements associated to the log statements. Let me know if this really needs to be changed to %s or if you are happy keeping them as .format()
Let me know what you think of the changes overall.

In longer term of things, I would keep the log configuration in a file somewhere, and have it loaded into the program for both recover-dump and for the unit tests. That would help protect the configuration from changes (as any changes would lead to unit test failure)

Please review and let me know what you think :) 
Again thanks for your feedback and reviews. Much appreciated

Patch-set: 3