Update patch set 3
Patch Set 3:

Hi there, apologies for the late follow-up -- just wondering -- you had mentioned we shouldn't forget to close the ticket -- I wonder if there's anything else that needs to be done? Have you had the time to review this? I think once this is merged, we can merge in T277160 with the rebase with logger

Patch-set: 3