Xtrabackup: Increase default open-files-limit to match production
S3 started to fail to prepare after it was migrated to buster.

Increase the maximum amount of open file descriptors to 200000,
the same as core production db instances, as that make it work
nicely for s3.

This is probably due to an optimization of InnoDB on later mariadb
versions, combined with some increase in s3 objects.

Bug: T281094
Change-Id: I0888845b9acfd08c8ca755d60e440ea39f8e8d36