Refactor check_mariadb_backups.py and add enough tests for it
check_mariadb_backups is the fundamental piece to detect errors
on database backups- despite having done in a rush.

Refactor so it has smaller, more logical methods, and also add
tests to make sure refactoring didn't break the logic, or later
enhancements, specially the alerting logic.

While the initial rationale for the change was avoiding hardcoding
the sections by reading that from the local file of section_ports
through wmfmariadbpy.dbutil, this was considered awkward at the
moment, and the infrastructure not ready, requiring an ugly patch,
so that work will be done in a future patch.

Improve also a bit the formatting of the output message, so errors
are easier to read on icinga, including backups sizes and
percentages.

Bug: T138562
Change-Id: Ie118920e55e2720de8e44567b41dd56e3c8df4d3