Check that xtrabackup --prepare is using the same version
When running the second phase of backup preparation for
mariabackup (xtrabackup --prepare), it is possible to
accidentally run it on a different version from that
of the backup source server.

While this should probably be fixed upsteam:
https://jira.mariadb.org/browse/MDEV-23718
Implement a failsafe on orchestration layer so that if this
case is detected, backup immediately fails.

This will prevent backup preparation weirdness, or, in a
worse case scenario, undetected corruption.

For now, the requirements have been set like this:
* Vendor (e.g. MariaDB) should be the same
* Major version (e.g. 10.4) should be the same
* xtrabackup minor should be the same or higher than server

But we can tune that in the future for more strict or looser
standards.

A few formatting improvements also applied, thanks to now
targetting bullseye as the main plattform (more modern python
version). Add more tests to make sure the refactoring is not
breaking the basic logic.

Bug: T253959
Change-Id: Id25fa4626bad86b7825dd4505a4985e975eb9617