check: Read list of valid sections/valid backup jobs from a file
In order to make easier to read the list of valid database section
names (or job names, as they tend to match for tracking, but don't
necessarily have to), read those from a local config file at:
/etc/wmfbackups/valid_sections.txt

This prevents accidentally checking a backup job that is not
intended (e.g. due to a spelling error).

This is intended as a better solution than the current
hardcoded system, but pending for a more global api/solution
at: T303178

The file format is simple: one line per section name in UTF-8
format, spaces and blank lines are ignored.
The file must contain at least one section to avoid errors.
Because config errors should be spotted easilly, they will
return an UKNOWN on icinga (and not the CRITICAL that other
fatal errors return).

I've also tuned the error code if there is an argument parsing
error to have the same behaviour.

The directory will be created by the debian package, but the
file will not be provided by it- it will be intended to be handled
at configuration management level (Puppet).

Bug: T301315
Change-Id: I8b20df1603651f3cbdeb48a3d789f59c23f6acf3