BackupStatistics: Change the statistics yaml file to a mysql config
The BackupStatistics used to parse a yaml file to connect to mysql.
Now, instead of handing mysql credentials as part of the config,
convert that into a .ini file, so it can be sent with full
flexibility to the mysql connector, and instead remove all the
mysql configuration options and merge them into one: stats_file,
with the patch of the mysql config file.

This will allow to handle how backups connect to mysql from
puppet, including tls options.

Update also remote-backup-mariadb to deprecate the old options
and update man pages and other help pages.

Release 0.8.3.

Bug: T351491